t/t-batch-*retries-ratelimit.sh: correct typos
Prior to updating the tests in our t/t-batch-retries-ratelimit.sh
and t/t-batch-storage-retries-ratelimit.sh test scripts, which we expect
to do in subsequent commits in this PR, we first correct the text of
the error messages output by several tests if the "git lfs clone"
command were to fail unexpectedly.

We also adjust the indentation of the lines which output these messages,
along with a few other lines in the same source files where we have
tab characters in place of space characters.