t/cmd/lfstest-gitserver.go: use api in retry keys
In commit 4fe0a8db4f94725078babc8eadc98f318655a05b of PR #3449 we
updated our lfstest-gitserver test utility program to simulate rate
limits on specific resources, allowing it to be used to validate the
changes made in that PR to respect 429 HTTP status code responses
sent by servers to the Git LFS client.

In the lfstest-gitserver utility, the checkRateLimit() function
is called when specific resources are requested by the client, and
it determines whether the test server should allow the request to
proceed as normal or whether a 429 response should be returned.

When each resource is first requested, a timer is enabled by the
checkRateLimit() function and the function returns values to the
caller indicating that a 429 response with a Retry-After header should
be sent to the client.  Once the timer has expired, the function
permits a fixed number of subsequent requests for the same resource
to proceed without delay, after which the timer is re-enabled and
future requests will again receive 429 responses.

To track the timer and request count values for each resource,
the checkRateLimit() function constructs a key from three of its
four arguments.  However, the function's first argument, named "api",
is at present left unused.

This oversight stems from the design of the similar
incrementRetriesFor(), which also constructs per-resource keys to
help track request counts.  These counts are used when validating the
retry behaviour of the Git LFS client after it receives repeated
5xx HTTP status codes while uploading or downloading objects.

In both functions, the "api" argument is unused in the construction
of per-resource key strings.  The argument is used by the
incrementRetriesFor() function to construct a value it checks
against the entry, if any, in the oidHandlers map for a given
object ID.  This map determines whether a specific object's content
(and therefore Git LFS SHA-256 object ID) indicates that the test
utility server should behave differently for that object.  In the
case of the incrementRetriesFor() function, this indicates whether
a retry count should be kept and reported for the requests for the
same object.

Arguably, we could simply drop the "api" argument from the set
of values passed to the checkRateLimit() function.  However, since
all our callers of both this function and the incrementRetriesFor()
function take care to pass an appropriate value for this argument,
either "batch" or "storage", we instead just incorporate the "api"
argument into the per-resource keys constructed by both functions,
and refactor this key generation into a small getResourceKey()
helper function.

This change should allow for the checkRateLimit() and
incrementRetriesFor() functions to be used more broadly in the
future, without concern that we may accidentally replicate the
same resource keys across tests of different APIs.