t: check batch API retries after 429s when cloning
The "batch clone causes retries" test was added to our
t/t-batch-retries-ratelimit.sh test script in commit
540d93a1ac56f3d9cf93519c7812ad55510d42cb of PR #3449, with the
intent that it should validate the Git LFS client's handling of 429
HTTP status code responses to individual object transfer requests
during clone operations.

Later, in commit 32f571d81d56c1f21977b67bcfd42a14f651c4e4 of PR #4573,
this test was revised to instead validate how the client handles
429 responses to Git LFS Batch API requests during clone operations.
In the same commit, the "batch clone with multiple files causes retries"
test was also added, which performs the same checks but when more than
one Git LFS object is being transferred.

Like the other tests in this test script, these tests depend on our
lfstest-gitserver utility program to track the requests made for
certain resources and generate 429 responses under specific conditions.

In particular, when a resource is first requested, a timer is enabled
by the checkRateLimit() function and the function returns values to
the caller indicating that a 429 response with a Retry-After header
should be sent to the client.  Once the timer has expired, the function
permits a fixed number of subsequent requests for the same resource
to proceed without delay, after which the timer is re-enabled and
future requests will again receive 429 responses.

In the case of the "batch clone causes retries" and "batch clone with
multiple files causes retries" tests, the tests clone a bare repository,
create Git LFS objects in it, and push that change to the test server.
The tests then clone the repository again, and expect that during this
second clone the server will apply its rate-limiting behaviour.  The
intent of the tests is that the Git LFS client should be forced to retry
its Batch API requests during this clone operation until it no longer
receives 429 responses from the test server.

However, in practice, the push operation that uploads the Git LFS objects
to the server is the one which is rate-limited and delayed by the test
server, because the checkRateLimit() function starts its per-resource
timer the first time it is invoked for a given resource.  (The resource
is defined by just the API name "batch" and the test repository name,
as the other elements of the resource key are left blank in the calls
to the function for these repositories.)  The second clone operation
then proceeds without delay and without receiving a 429 response.

To resolve this problem with the tests, we add a /limits/ endpoint
to our lfstest-gitserver program which accepts query arguments
defining a given resource key, and a value to set as that key's
remaining number of non-rate-limited requests.  If the value provided
is the string "max", the default maximum value as defined by the
refillTokenCount constant is used.

Next, we define a set_server_rate_limit() test helper function in
our t/testhelpers.sh shell library, which makes a GET request to our
new /limits/ endpoint and passes its input arguments as query
parameters.

We can then invoke this helper function twice in our "batch clone
causes retries" and "batch clone with multiple files causes retries"
tests, the first time before pushing the commit with the new Git LFS
objects, and the second time afterwards.  For the first call we provide
a "max" value for the number of available non-rate-limited requests,
ensuring the push proceeds without delay, and for the second call we
reset the number of available non-rate-limited requests to zero, so
that the clone operation will now receive a 429 response to its Git
LFS Batch API request and be forced to wait and retry that request.

Finally, this allows us to check the number of retries performed by
the client during the second clone operation.  To do this we utilize
the same approach we introduced in a prior commit in this PR for
several other tests in the same test script, where we check the
number of times the "enqueue retry" trace log message is output.
For the test with multiple Git LFS objects, we confirm that we see
this message exactly once for each object, and only with the "#1"
retry count value.

These checks requires that we set the GIT_TRACE environment variable
for the "git lfs clone" commands and capture the commands' output
into clone.log files.  Because we use shell pipelines and the tee(1)
command for this purpose, we also have to adjust how we check the
exit code from the "git lfs clone" commands since they are the first
command in the pipelines.