t/t-batch-*retries-ratelimit.sh: test retry counts
In a subsequent commit in this PR we will resolve a problem whereby
we do not honour the delay specified in a Retry-After header when
it is received along with a 429 HTTP status code in response to a
request to upload or download an individual Git LFS object.

As part of this change, we will update several of the relevant tests
in our t/t-batch-storage-retries-ratelimit.sh test script to confirm
that we only retry our requests to the server after the interval
specified by the Retry-After header has elapsed.

Prior to making these changes, we first update two other tests
in the t/t-batch-storage-retries-ratelimit.sh test script to check
that multiple retry attempts are made when no Retry-After header is
sent with a 429 response to an object upload or download request.
This behaviour from the Git LFS client is correct, as we expect it
to use an exponential backoff retry strategy under these conditions,
a feature that was implemented in PR #4097.

The checks we add to the "batch storage upload causes retries (missing
header)" and "batch storage download causes retries (missing header)"
tests verify that we see multiple "retrying object" and "enqueue retry"
trace log messages during their "git lfs push" and "git lfs pull"
commands, respectively.  Because only a single Git LFS object is
transferred to the server by either of these tests, it is sufficient
for us to check that the "enqueue retry" messages appear with both
"#1" and "#2" retry counts to confirm that all these represent retries
for that single object.

We cannot yet checks like these to the other tests in the
t/t-batch-storage-retries-ratelimit.sh test script, though, until we
resolve the problem described above, namely that at present we fail to
honour the delay specified in a Retry-After header received in response
to an individual Git LFS object transfer request.

We can, however, add checks of the number of retry attempts to some
of the tests in the t/t-batch-retries-ratelimit.sh test script, as
these tests validate the retry attempts made when a 429 response is
received to a Git LFS Batch API request, not an individual object
transfer request.

Because the Git LFS client honours the delay specified in a Retry-After
header sent a 429 response to a Batch API request, we can verify that
only a single retry is performed for each object (since the specified
interval has elapsed) in this case, so we add checks of this behaviour
to our "batch upload causes retries" and "batch upload with multiple
files causes retries" tests.  For the latter test, we verify that
although we see two "enqueue retry" trace log messages, they both have
the retry count "#1" and so must apply to different Git LFS objects,
as we expect.

We can also confirm that multiple retries for a single object are
performed when a Batch API request receives a 429 response without a
Retry-After header, so we add checks of this behaviour to our "batch
upload causes retries (missing header)" test.  In this instance we
expect the same exponential backoff retry strategy implemented in
PR #4097 to be used by the client, and use the same type of checks
we used in the "(missing header)" tests in the
t/t-batch-storage-retries-ratelimit.sh test script.