.github/workflows: upgrade to Go 1.24
Go version 1.24 has been released, and because we aim to build and test
Git LFS against only supported versions of Go, we upgrade our GitHub
Actions CI workflows to test against Go versions 1.24 and 1.23.

This resolves a problem now seen in our "Build with specific Go" CI job,
where we install the latest version of the "goimports" package and it
fails because the "x/tools" module requires Go 1.23 as of commit
golang/tools@94db7107945332a789135d458a0f0de1d7c00ddb, and we are still
using Go 1.22 for that CI job.