t: skip file deletion in missing object push tests
Our t/t-pre-push.sh and t/t-push-failures-local.sh test scripts include
a number of tests which validate the behaviour of the "git lfs pre-push"
and "git push" commands when Git LFS objects are missing or corrupt.

In a subsequent commit in this PR we expect to update the Git LFS
client to remove some non-functional code which attempts to recreate
missing Git LFS objects during push operations.  In many cases this
change will cause the client to report missing objects in an earlier
stage of push operations than it does now.  We also expect to reword
the error message the client will output in such cases.

Before we make these changes, we update two of the tests where the
"lfs.allowIncompletePush" Git configuration option is set to "true"
so that although they remove one Git LFS object file from the local
storage directories under .git/lfs/objects, they do not also remove
the corresponding Git LFS pointer from the repository's most recent
commit, or remove the copy of the object file's contents from the
current working tree.

These additional removal actions are not necessary to simulate the
test condition of a Git LFS object file missing from the local
storage directories, and so long as the "lfs.allowIncompletePush"
option is set to "true", have no effect on the outcome of the
"git lfs pre-push" and "git push" commands.

Note that several other related tests in the t/t-pre-push.sh and
t/t-push-failures-local.sh scripts also perform the same action
of removing an object file from the local storage directories and then
also deleting the associated Git LFS pointer and corresponding file
in the working tree.  In a subsequent commit in this PR we will
eliminate these steps from the other tests as well.

However, we first need to address the fact that the Git LFS client
outputs different error messages when an object file is missing,
depending on whether or not a file exists in the working tree at
the path associated with the Git LFS pointer corresponding to the
object.  The file in the working tree may have any content, or be
entirely empty; its simple presence is enough to change the error
message output by the Git LFS client and the point at which the
client abandons the upload push operation, so long as the
"lfs.allowIncompletePush" configuration option is set to its
default values of "false".

In a subsequent commit in this PR we will remove the code which
causes this variable behaviour, which will simplify our handling of
missing objects during upload transfer operations.

At the same time we will also update the relevant tests in the
t/t-pre-push.sh and t/t-push-failures-local.sh scripts, and in
particular we will adjust them so that after they remove a Git LFS
object file from the local storage directories, they do not also
remove the corresponding Git LFS pointer from the repository's Git
history or the associated copy of the object file's contents in the
current working tree.  Thus these tests' setup steps will again
exactly match those of the tests we have modified in this commit.